git-svn-id: https://subversion.assembla.com/svn/kbasm/cpgf/trunk@121 5ea4cdcf-77fc-412b-903d-47c6f9a32d48

+++ src/gmetaclass.cpp
@@ -261,7 +261,7 @@ size_t GMetaClass::getTypeSize() const
 	return this->baseData->getObjectSize();
 }
 
-const GMetaConstructor * GMetaClass::getConstructorByParamCount(unsigned int paramCount) const
+const GMetaConstructor * GMetaClass::getConstructorByParamCount(size_t paramCount) const
 {
 	this->ensureRegistered();
